Integrate proven behaviour into auth command
