+++ notion-console/src/commonMain/kotlin/io/ashdavies/notion/AuthCommand.kt
@@ -1,9 +1,10 @@
 package io.ashdavies.notion
 
 import androidx.compose.runtime.Composable
-import io.ashdavies.playground.Token
 import io.ashdavies.playground.TokenQueries
 import kotlinx.cli.ExperimentalCli
+import java.awt.Desktop
+import java.net.URI
 
 @Composable
 @ExperimentalCli
@@ -20,12 +21,11 @@ internal fun AuthCommand(
         onExecute = {
             val token = tokenQueries.select().executeAsOneOrNull()
             if (token != null) {
-                onAuthState(AuthState.Authenticated(token))
+                onAuthState(AuthState.Authenticated)
             } else {
                 onAuthState(AuthState.Awaiting("http://localhost:8080/callback"))
-                val result = Token(getAccessToken(OAuthProvider.Notion))
-                onAuthState(AuthState.Authenticated(result))
-                tokenQueries.insert(result)
+                getAccessToken { Desktop.getDesktop().browse(URI(it)) }
+                onAuthState(AuthState.Authenticated)
             }
         },
         name = "login",
@@ -40,13 +40,5 @@ internal fun AuthCommand(
 
 internal sealed interface AuthState : NotionState {
     data class Awaiting(val userPromptUri: String) : AuthState
-    data class Authenticated(val token: Token) : AuthState
+    object Authenticated : AuthState
 }
-
-private fun Token(value: AccessToken) = Token(
-    workspaceIcon = value.firstOrThrow("workspaceIcon"),
-    workspaceName = value.firstOrThrow("workspaceName"),
-    workspaceId = value.firstOrThrow("workspaceId"),
-    botId = value.firstOrThrow("botId"),
-    accessToken = value.accessToken,
-)
