--- notion-console/src/commonMain/kotlin/io/ashdavies/notion/OAuthProvider.kt
@@ -1,6 +1,0 @@
-package io.ashdavies.notion
-
-public sealed interface OAuthProvider {
-    public object Google : OAuthProvider
-    public object Notion : OAuthProvider
-}
