--- notion-console/src/commonMain/kotlin/io/ashdavies/notion/OAuthSettings.kt
@@ -1,10 +1,0 @@
-package io.ashdavies.notion
-
-import io.ktor.server.auth.OAuthServerSettings
-
-public object OAuthSettings : (OAuthProvider) -> OAuthServerSettings {
-    override fun invoke(provider: OAuthProvider): OAuthServerSettings = when (provider) {
-        is OAuthProvider.Google -> GoogleSettings()
-        is OAuthProvider.Notion -> NotionSettings()
-    }
-}
